requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
Jinja2==3.1.4
psycopg2-binary==2.9.9
cryptography==42.0.8
python-multipart==0.0.9
Pillow==11.2.1
